update for new snapcraft, fix old bugs
Migrated from ci-management: https://github.com/edgexfoundry/ci-management/pull/617
Changes by Ian Johnson <ian.johnson@canonical.com>

Due to some libraries that snapcraft/python use for outputting, the ordering of
some output gets lost when it is piped through docker. To fix this we can just
run snapcraft through unbuffer, which is provided through the expect package.

This is needed because now snapcraft will use "snap pack" from the host, and so
we need the snap command in the docker container alongside the snapcraft snap.

The previous behavior would not actually fail the build setup if for example the
SHA's didn't match, but now if the SHA's don't match, sha512sum will exit
non-zero and using && results in the whole command failing.

We also should delete the sha512 file as well, it is not needed during the
build.

Fixes: #517

Signed-off-by: Ernesto Ojeda <ernesto.ojeda@intel.com>