feat: replace BusyBox mktemp with GNU mktemp
In the process of attempting to rebuild this image, it seems that alpine v2.6 is no longer available anywhere so I had to upgrade sysstat to a newer version. I know the version mismatch between the host and docker container causes issues. However, I found sar files can be converted to different version formats. I have documented this in the Dockerfile.logs-publish.

Signed-off-by: Ernesto Ojeda <ernesto.ojeda@intel.com>